services/shared/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
boto3==1.34.0
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
razorpay==1.4.1
cryptography==41.0.7
httpx==0.25.2

